Add AlignToAngle
A new state the aligns the robot to an angle within 1 degree of gyro
reading, and has a timeout condition equal to the powerValue. This
assignment can be done because the power is set based on a p-style power
setting, that sets power based on the gyro's offset from the goal.