Input Reading!
This file can read a text file stored on the phone, and will run code
based off of it. This lets us "change code in queueing", and gives us
lots of flexibility. The biggest todo for this project is to make it
more user friendly, and easier to write code quickly.